requirements.txt
autopep8==1.6.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
google==3.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
python-dotenv==0.19.2
requests==2.25.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
